Debugged & fixed plane controls, added rotation to propeller
Fixed issue with planes movement updating the translate to use .forward instead of .back to allow the plane to go forward. Updated plane tilt rotation use vertical input to allow the player to control the planes up and down direction. Created an object reference for the propeller on the player controller class for the plane rotating the propeller on the Z axis.